% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{customModalDialog}
\alias{customModalDialog}
\title{This is just shiny modalDialog with one extra line to add a button in}
\usage{
customModalDialog(
  ...,
  input,
  session,
  displayXClose = TRUE,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l", "xl", "fs"),
  easyClose = FALSE,
  fade = TRUE
)
}
\arguments{
\item{...}{UI elements for the body of the modal dialog box.}

\item{input}{The shin input object}

\item{session}{This is added to carry namespace to the UI if override not used.}

\item{displayXClose}{Logical Whether to disable the extra close `X` button in upper right.
TRUE = Enabled, FALSE = disabled}

\item{title}{An optional title for the dialog.}

\item{footer}{UI for footer. Use NULL for no footer.}

\item{size}{One of "s" for small, "m" (the default) for medium, or "l" for large.}

\item{easyClose}{If TRUE, the modal dialog can be dismissed by clicking outside the dialog box,
or be pressing the Escape key. If FALSE (the default), the modal dialog can't be dismissed in those ways;
instead it must be dismissed by clicking on a modalButton(), or from a call to removeModal() on the server.}

\item{fade}{If FALSE, the modal dialog will have no fade-in animation (it will simply appear rather than fade in to view).}
}
\description{
The shiny modalDialog function with added upper right close button for bootstrap 5
}
