% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dbUpdate}
\alias{dbUpdate}
\title{dbUpdate}
\usage{
dbUpdate(con = NULL, from, to, data = LOCAL)
}
\arguments{
\item{con}{database connection that exists in the parent. Placeholder arg.}

\item{from}{DF of data going to data base}

\item{to}{The database table name where the update should happen. DB table should match the DF in `from`}
}
\description{
Case conditions and set up to guide insert or
update of the database tables. Guides the delete, upsert... functions.
}
