% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_auth0.R
\name{run_app_auth0}
\alias{run_app_auth0}
\title{run_app_auth0}
\usage{
run_app_auth0(...)
}
\arguments{
\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
run_app_auth0
}
